ocr-service/requirements.txt
flask==3.0.3
pytesseract==0.3.13
opencv-python==4.10.0.84
pdf2image==1.17.0
numpy==2.1.1
Pillow==10.4.0
